requirements.txt
mysql-connector-python==8.0.32 ; python_version >= "3.10" and python_version < "4.0"
protobuf==3.20.3 ; python_version >= "3.10" and python_version < "4.0"
